pyproject.toml
[project]
name = "chembed-res"
version = "0.1.0"
description = "MTEB evaluation scripts for ChemRxiv and ChemTEB benchmarks"
requires-python = ">=3.12"
dependencies = [
    "mteb==2.7.1",
    "sentence-transformers>=5.2.0",
    "einops>=0.8.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "boto3>=1.42.31",
    "openai>=2.15.0",
    "huggingface-hub>=0.34.0,<1.0",
    "ruff>=0.14.13",
]

[tool.uv]
package = false
